refactor: interceptor allow uri 명시적으로 변경
